Address PR suggestions and fix the dtype in 'random' init

+++ sklearn/mixture/_base.py
@@ -119,8 +119,10 @@ class BaseMixture(DensityMixin, BaseEstimator, metaclass=ABCMeta):
             )
             resp[np.arange(n_samples), label] = 1
         elif self.init_params == "random":
-            resp = random_state.uniform(size=(n_samples, self.n_components))
-            resp /= resp.sum(axis=1)[:, np.newaxis].astype(X.dtype)
+            resp = random_state.uniform(size=(n_samples, self.n_components)).astype(
+                X.dtype
+            )
+            resp /= resp.sum(axis=1)[:, np.newaxis]
         elif self.init_params == "random_from_data":
             resp = np.zeros((n_samples, self.n_components), dtype=X.dtype)
             indices = random_state.choice(
